Frontend finalizado, backend finalizado, ejb corrigido mas com bug

+++ backend-module/src/main/java/com/example/backend/src/main/java/com/bip/DTO/BeneficioCadastroDTO.java
@@ -1,12 +1,14 @@
 package com.bip.DTO;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class BeneficioCadastroDTO {
 	private String nome;
 	private String descricao;
-	private float valor;
+	private BigDecimal valor;
 	private boolean ativo;
 
 }
