+++ backend-module/src/main/java/com/example/backend/src/main/java/com/bip/model/Beneficio.java
@@ -1,5 +1,7 @@
 package com.bip.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -22,7 +24,7 @@ public class Beneficio {
 	private String descricao;
 
 	@Column
-	private float valor;
+	private BigDecimal valor;
 
 	@Column
 	private boolean ativo;
